refactor: Abstract interval unit and value logic
