+++ src/locale/en.json
@@ -6,9 +6,9 @@
     "CUSTOM_INTERVAL_SET_FAILURE": "Oops, your custom interval of %{refreshInterval} minutes must be between %{refreshIntervalMinimum} and %{refreshIntervalMaximum} minutes. Please try again.",
     "SITE_DEBUG_MODE_FAILURE": "There was an issue updating settings for debug mode. Please try again.",
     "SITE_REFRESHED_FAILURE": "There was an issue refreshing your site. Please try again.",
-    "SITE_REFRESHED_SUCCESS": "You've successfully refreshed your site. All connected browsers will refresh within %{refreshInterval} seconds.",
+    "SITE_REFRESHED_SUCCESS": "You've successfully refreshed your site. All connected browsers will refresh within %{refreshIntervalValue} %{refreshIntervalUnit}.",
     "PAGE_REFRESHED_FAILURE": "There was an issue refreshing your page. Please try again.",
-    "PAGE_REFRESHED_SUCCESS": "You've successfully refreshed your page. All connected browsers will refresh within %{refreshInterval} seconds.",
+    "PAGE_REFRESHED_SUCCESS": "You've successfully refreshed your page. All connected browsers will refresh within %{refreshIntervalValue} %{refreshIntervalUnit}.",
     "SITE_SETTINGS_UPDATED_FAILURE": "There was an issue updating your settings. Please try again.",
     "SITE_SETTINGS_UPDATED_SUCCESS": "You've successfully updated settings for Force Refresh."
   },
